Add `parse_period_id_to_month()` as private helper
Add the function to parse period IDs into month numbers as
a private helper. This is the more appropriate place to define
it, as opposed to the script to build and update the internal
data as it used to.

`bps_add_date()` now throws an error when there is no period
table in the internal data. This gives a more helpful context
to users.